Refactor loading ontology in a separate function
This patch refactors many functions from `render` module to
`intermediate` module so that the downstream users can load the ontology
and do analysis on top of it without having to render the scenarios to
HTML.
